requirements.txt
fastapi==0.115.0
sqlalchemy==2.0.30
pydantic==2.9.2
requests==2.32.3
psycopg2-binary==2.9.9
faker==40.5.1
pandas
typing
email-validator
numpy
scikit-learn
seaborn
tensorflow